requirements.txt
click==8.1.2
pytsk3==20230125
PyYAML==6.0
python-evtx==0.7.4